@@ -5,75 +5,76 @@
   "main": "main.js",
   "private": true,
   "scripts": {
-    "postinstall": "npx electron-builder install-app-deps",
+    "postinstall": "npm run postinstall:electron && electron-builder install-app-deps",
+    "postinstall:web": "node postinstall-web",
+    "postinstall:electron": "node postinstall",
     "ng": "ng",
-    "start": "npm-run-all -p ng:serve electron:serve",
-    "lint": "ng lint",
-    "build": "ng build && npm run electron:tsc",
-    "build:prod": "ng build --prod && npm run electron:tsc",
-    "ng:serve": "ng serve -o",
-    "electron:tsc": "tsc main.ts",
-    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:tsc && electron . --serve",
+    "start": "npm run postinstall:electron && npm-run-all -p ng:serve electron:serve",
+    "build": "npm run postinstall:electron && npm run electron:serve-tsc && ng build",
+    "build:dev": "npm run build -- -c dev",
+    "build:prod": "npm run build -- -c production",
+    "ng:serve": "ng serve",
+    "ng:serve:web": "npm run postinstall:web && ng serve -o",
+    "electron:serve-tsc": "tsc -p tsconfig-serve.json",
+    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:serve-tsc && electron . --serve",
     "electron:local": "npm run build:prod && electron .",
-    "electron:linux": "npm run build:prod && npx electron-builder build --linux",
-    "electron:windows": "npm run build:prod && npx electron-builder build --windows",
-    "electron:mac": "npm run build:prod && npx electron-builder build --mac",
-    "test": "karma start ./karma.conf.js",
-    "pree2e:build": "webdriver-manager update --gecko false && npm run ng serve",
-    "pree2e:protractor": "wait-on http-get://localhost:4200/ && protractor ./protractor.conf.js",
-    "e2e": "npm-run-all -p -r pree2e:build pree2e:protractor"
+    "electron:linux": "npm run build:prod && electron-builder build --linux",
+    "electron:windows": "npm run build:prod && electron-builder build --windows",
+    "electron:mac": "npm run build:prod && electron-builder build --mac",
+    "test": "npm run postinstall:web && ng test",
+    "e2e": "npm run build:prod && mocha --timeout 300000 --require ts-node/register e2e/**/*.spec.ts",
+    "lint": "ng lint"
   },
   "devDependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "5.2.3",
-    "@angular/cli": "1.7.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/language-service": "5.2.9",
-    "@angular/material": "5.2.3",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "@types/core-js": "0.9.36",
-    "@types/jasmine": "2.8.6",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "7.0.7",
-    "angular-expressions": "0.3.0",
-    "codelyzer": "4.2.1",
-    "core-js": "2.5.5",
-    "cross-env": "5.1.4",
-    "docxtemplater": "3.6.2",
-    "dotenv": "5.0.1",
-    "electron": "1.8.4",
-    "electron-builder": "20.8.1",
-    "electron-reload": "1.2.2",
-    "file-saver": "1.3.8",
-    "fs-extra": "5.0.0",
-    "jasmine-core": "2.99.1",
+    "@angular-devkit/build-angular": "0.803.6",
+    "@angular/animations": "8.2.8",
+    "@angular/cdk": "8.2.2",
+    "@angular/cli": "8.3.6",
+    "@angular/common": "8.2.8",
+    "@angular/compiler": "8.2.8",
+    "@angular/compiler-cli": "8.2.8",
+    "@angular/core": "8.2.8",
+    "@angular/forms": "8.2.8",
+    "@angular/language-service": "8.2.8",
+    "@angular/material": "8.2.2",
+    "@angular/platform-browser": "8.2.8",
+    "@angular/platform-browser-dynamic": "8.2.8",
+    "@angular/router": "8.2.8",
+    "@types/jasmine": "3.3.16",
+    "@types/jasminewd2": "2.0.6",
+    "@types/jszip": "3.1.6",
+    "@types/mocha": "5.2.7",
+    "@types/node": "12.6.8",
+    "angular-expressions": "1.0.0",
+    "chai": "4.2.0",
+    "codelyzer": "5.1.0",
+    "core-js": "3.1.4",
+    "docxtemplater": "3.14.3",
+    "electron": "6.0.10",
+    "electron-builder": "21.2.0",
+    "electron-reload": "1.5.0",
+    "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "jszip": "2.6.1",
-    "jszip-utils": "0.0.2",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "mdi": "2.0.46",
-    "npm-run-all": "4.1.2",
-    "npx": "10.0.1",
-    "protractor": "5.3.1",
-    "replace": "0.3.0",
-    "roboto-fontface": "0.8.0",
-    "rxjs": "5.5.8",
-    "ts-node": "4.1.0",
-    "tslint": "5.9.1",
-    "typescript": "2.6.2",
-    "wait-on": "2.1.0",
-    "webdriver-manager": "12.0.6",
-    "zone.js": "0.8.26"
+    "jszip": "3.2.2",
+    "karma": "4.2.0",
+    "karma-chrome-launcher": "3.0.0",
+    "karma-coverage-istanbul-reporter": "2.1.0",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.4.2",
+    "mdi": "2.2.43",
+    "mocha": "6.2.0",
+    "npm-run-all": "4.1.5",
+    "roboto-fontface": "0.10.0",
+    "rxjs": "6.5.3",
+    "spectron": "8.0.0",
+    "ts-node": "8.3.0",
+    "tslint": "5.18.0",
+    "typescript": "3.5.3",
+    "wait-on": "3.3.0",
+    "webdriver-manager": "12.1.5",
+    "zone.js": "0.9.1"
+  },
+  "engines": {
+    "node": ">=10.9.0"
   }
 }
