@@ -1,31 +1,28 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy } from '@angular/core';
-import { animate, state, style, transition, trigger } from '@angular/animations';
+import { animate, AnimationEvent, state, style, transition, trigger } from '@angular/animations';
 import { Portal } from '@angular/cdk/portal';
 
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
+import { Observable, Subject } from 'rxjs';
 
 export type TooltipVisibility = 'initial' | 'visible' | 'hidden';
 
-
 @Component({
-  selector: 'lms-dynamic-content-tooltip',
+  selector: 'app-dynamic-content-tooltip',
   templateUrl: './dynamic-content-tooltip.component.html',
   styleUrls: ['./dynamic-content-tooltip.component.scss'],
-  preserveWhitespaces: false,
   animations: [
     trigger('state', [
       state('initial, void, hidden', style({transform: 'scale(0)'})),
       state('visible', style({transform: 'scale(1)'})),
       transition('* => visible', animate('150ms cubic-bezier(0.0, 0.0, 0.2, 1)')),
       transition('* => hidden', animate('150ms cubic-bezier(0.4, 0.0, 1, 1)')),
-    ])
+    ]),
   ],
   host: {
     '[style.zoom]': 'visibility === "visible" ? 1 : null',
-    '(body:click)': 'this.handleBodyInteraction()'
+    '(body:click)': 'this.handleBodyInteraction()',
   },
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DynamicContentTooltipComponent implements OnDestroy {
   public contentPortal: Portal<Object>;
@@ -70,13 +67,12 @@ export class DynamicContentTooltipComponent implements OnDestroy {
     return this._onHide.asObservable();
   }
 
-
   public animationStart(): void {
     this._closeOnInteraction = false;
   }
 
   public animationDone(event: AnimationEvent): void {
-    const toState: TooltipVisibility = event['toState'] as TooltipVisibility;
+    const toState: TooltipVisibility = event.toState as TooltipVisibility;
 
     if (toState === 'hidden' && !this.isVisible()) {
       this._onHide.next();
