@@ -4,11 +4,10 @@ import { Operand } from '../../expression/operand';
 import { DisjunctiveExpression } from '../../expression/disjunctive-expression';
 import { Expression } from '../../expression/expression';
 
-
 @Component({
   selector: 'app-discrete-expression',
   templateUrl: './discrete-expression.component.html',
-  host: { class: 'component-wrapper' }
+  host: { class: 'component-wrapper' },
 })
 export class DiscreteExpressionComponent {
 
@@ -16,7 +15,7 @@ export class DiscreteExpressionComponent {
   @Input() public function: Expression;
   @Input() public nested: boolean = false;
 
-  public isOperand(operand): operand is Operand {
+  public isOperand(operand: Operand | Expression): operand is Operand {
     return operand instanceof Operand;
   }
 
