@@ -1,15 +1,13 @@
 import { enableProdMode, ViewEncapsulation } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { RootModule } from './app/root/root.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule, {
-    preserveWhitespaces: false,
-    defaultEncapsulation: ViewEncapsulation.None
-  })
+platformBrowserDynamic()
+  .bootstrapModule(RootModule, { defaultEncapsulation: ViewEncapsulation.None })
   .catch(err => console.error(err));
